+++ src/MathLib/Matrix.cpp
@@ -60,7 +60,7 @@ matrix_t* matrix_scalar_set(matrix_t* a, double h) {
         return nullptr;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] = h;
     }
     return a;
@@ -72,7 +72,7 @@ matrix_t* matrix_scalar_add(matrix_t* a, double h) {
         return a;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] += h;
     }
     return a;
@@ -84,7 +84,7 @@ matrix_t* matrix_scalar_mul(matrix_t* a, double h) {
         return a;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] *= h;
     }
     return a;
@@ -104,7 +104,7 @@ matrix_t* matrix_add(matrix_t* a, matrix_t* b) {
         return a;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] += b->data[idx];
     }
     return a;
@@ -124,7 +124,7 @@ matrix_t* matrix_sub(matrix_t* a, matrix_t* b) {
         return a;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] -= b->data[idx];
     }
     return a;
@@ -136,7 +136,7 @@ matrix_t* matrix_heaviside(matrix_t* a) {
         return a;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] = (a->data[idx] > 0.0) ? 1.0 : 0.0;
     }
     return a;
@@ -148,7 +148,7 @@ matrix_t* matrix_random_f(matrix_t* a) {
         return a;
     }
 #pragma omp parallel for
-    for (int idx = 0; idx < a->rows * a->cols; idx++) {
+    for (int idx = 0; idx < static_cast<int>(a->rows * a->cols); idx++) {
         a->data[idx] = drand48();
     }
     return a;
